// src/lib/models/CertificateData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CertificateData = {
    cert_pem?: string;
    chain_pem?: string;
    domainName?: Array<string>;
    fullchain_pem?: string;
    priv_key?: string;
    s3DownloadUrl?: string;
    zipDir?: string;
};